refactor(client): 优化多部分预览不透明度及影响范围渲染逻辑
- 将配置中的多部分预览不透明度类型由float改为double，增强精度
- 在HudTooltipManager中新增解析作用范围方法，提升代码复用性
- 修改多处多部分预览相关代码，确保不透明度转换为float后再使用
- RenderEventListener新增影响范围持久化缓存，避免频繁计算和渲染
- 在渲染事件中根据游戏时间管理作用范围的显示时长，实现边框持续显示效果
- 优化智能放置器和结构扫描器的范围渲染，改用统一的颜色和持久化逻辑
- 精简渲染流程，添加null安全检查，提升稳定性和性能

+++ src/main/java/dev/dubhe/anvilcraft/client/event/LargeBlockPlacePreviewEventListener.java
@@ -260,7 +260,7 @@ public class LargeBlockPlacePreviewEventListener {
             expandRenderEntriesForGhost();
         }
         RenderType renderType = ModRenderTypes.BEACON_GLASS;
-        float alpha = AnvilCraftClient.CONFIG.multiPartPreviewGhostOpacity;
+        float alpha = (float) AnvilCraftClient.CONFIG.multiPartPreviewGhostOpacity;
         int color = boundColor;
         float red = FastColor.ARGB32.red(color) / 255f;
         float green = FastColor.ARGB32.green(color) / 255f;
@@ -335,7 +335,7 @@ public class LargeBlockPlacePreviewEventListener {
                     selectionPart.apply(poseStack);
                     OutlineRenderer.render(poseStack, vertexConsumer,
                         CubeSelection.outlines().get(selectionPart.geometry()), 1.0f, 1.0f, 1.0f,
-                        AnvilCraftClient.CONFIG.multiPartPreviewOutlineOpacity);
+                        (float) AnvilCraftClient.CONFIG.multiPartPreviewOutlineOpacity);
                     poseStack.popPose();
                 }
                 poseStack.popPose();
@@ -370,7 +370,7 @@ public class LargeBlockPlacePreviewEventListener {
                     poseStack.scale(1.001f, 1.001f, 1.001f);
                     BlockState partState = amplifier.placedState(part, state);
                     renderPart(poseStack, bufferSource, renderType, partState,
-                        AnvilCraftClient.CONFIG.multiPartPreviewGhostOpacity, 1.0f, 1.0f, 1.0f);
+                        (float) AnvilCraftClient.CONFIG.multiPartPreviewGhostOpacity, 1.0f, 1.0f, 1.0f);
                     poseStack.popPose();
                 }
             }
@@ -407,7 +407,7 @@ public class LargeBlockPlacePreviewEventListener {
             part.apply(poseStack);
             OutlineRenderer.render(poseStack, vertexConsumer,
                 CubeSelection.outlines().get(part.geometry()), red, green, blue,
-                AnvilCraftClient.CONFIG.multiPartPreviewOutlineOpacity);
+                (float) AnvilCraftClient.CONFIG.multiPartPreviewOutlineOpacity);
             poseStack.popPose();
         }
         poseStack.popPose();
